// server/src/core/dtos/customer/customer-basic.dto.ts
export class CustomerBasicDto {
    id?: number;
    customerRoleId: number;
    username: string;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    createdAt: Date;
    updatedAt: Date;
}
// DB
// id?: number;
// customer_role_id: string;
// username: string;
// first_name: string;
// last_name: string;
// email: string;
// password: string;
// phone: string;
// created_timestamp: Date;
// updated_timestamp: Date;